WIP: progress

+++ upload/php-s3-server-master/ffmpegProgress.php
@@ -1,7 +1,15 @@
 <?php
-	$progress = file_get_contents('./progress/'.$_GET['pid'].'.txt');
-	if ($progress == 100) {
-		unlink('./progress/'.$_GET['pid'].'.txt');
+	if (file_exists('./progress/'.$_GET['pid'].'.txt')) {
+		$progress = file_get_contents('./progress/'.$_GET['pid'].'.txt');
+		if ($progress == 100) {
+			unlink('./progress/'.$_GET['pid'].'.txt');
+			echo ('100');
+		}
+		else {
+			echo ($progress);
+		}
 	}
-	echo ($progress);
+	else {
+		echo ('0');
+	}
 ?>
